Issue #SC-211 feat:Renaming the file
